udpate api

+++ ManagementSystem.Common/Entities/Accountings/InventoryVoucherDetail.cs
@@ -7,11 +7,11 @@ namespace ManagementSystem.Common.Entities
         [ForeignKey("InventoryVoucher")]
         public int DocummentNumber { get; set; }
         public int ProductId { get; set; }
-        public int? Quantity { get; set; }
+        public float? Quantity { get; set; }
         public int? Price { get; set; }
         public int? DebitAccount { get; set; }
         public int? CreditAccount { get; set; }
-        public int? TotalMoneyBeforeTax { get; set; }
+        public float? TotalMoneyBeforeTax { get; set; }
         public int? DebitAccountMoney { get; set; }
         public int? CreditAccountMoney { get; set; }
         public int? PaymentDiscountAccount { get; set; }
